Try to debug UCB Increase iteration count Fix a bug

+++ 1 - Learning for Dynamic Pricing/UCBLearner.py
@@ -24,12 +24,9 @@ class UCBLearner(BanditLearner):
             for customer in last_customers:
                 reward = self._get_reward(customer, product_id)
                 self.rewards_per_arm_per_product[product_id][selected_price_index].append(reward)
-            self.means[product_id][selected_price_index] = int( # TODO: Why cast to int? There is something fishy or?
-                np.mean(self.rewards_per_arm_per_product[product_id][selected_price_index]))
+            self.means[product_id][selected_price_index] = np.mean(self.rewards_per_arm_per_product[product_id][selected_price_index])
             n = len(self.rewards_per_arm_per_product[product_id][selected_price_index])
             if n > 0:
                 self.widths[product_id][selected_price_index] = np.sqrt(2 * np.log(t + 1) / n)
             else:
                 self.widths[product_id][selected_price_index] = np.inf
-
-
